Rearrange HTMLKit internals
