+++ Sources/HTMLKit/Internal/HTMLBuilder.swift
@@ -1,0 +1,7 @@
+@_functionBuilder
+public class HTMLBuilder {
+
+    public static func buildBlock(_ children: HTML...) -> HTML {
+        return children
+    }
+}
