+++ Sources/HTMLKit/Internal/Variables/ContextualVariable.swift
